Refactor clients into separate modules
